backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
numpy>=2.1.0
pandas>=2.2.2
scikit-learn>=1.5.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.4.0
xgboost>=2.1.0
psycopg2-binary>=2.9.9
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic>=2.7.1
aiofiles==23.2.1
